Clean workspace and load all packages

load_all loads a package. It roughly simulates what happens when a package is installed and loaded with library().

load_clean_all(...)

Arguments

...
Additional arguments passed to pkgload::load_all().

Details

Currently load_all:

Loads all data files in data/. See load_data() for more details.

Sources all R files in the R directory, storing results in environment that behaves like a regular package namespace. See below and load_code() for more details.

Compiles any C, C++, or Fortran code in the src/ directory and connects the generated DLL into R. See pkgload::compile_dll() for more details.

Runs .onAttach(), .onLoad() and .onUnload() functions at the correct times.

If you use testthat, will load all test helpers so you can access them interactively. devtools sets the DEVTOOLS_LOAD environment variable to "true" to let you check whether the helpers are run during package loading.

Namespaces

The namespace environment <namespace:pkgname>, is a child of the imports environment, which has the name attribute imports:pkgname. It is in turn is a child of <namespace:base>, which is a child of the global environment. (There is also a copy of the base namespace that is a child of the empty environment.)

The package environment <package:pkgname> is an ancestor of the global environment. Normally when loading a package, the objects listed as exports in the NAMESPACE file are copied from the namespace to the package environment. However, load_all by default will copy all objects (not just the ones listed as exports) to the package environment. This is useful during development because it makes all objects easy to access.

To export only the objects listed as exports, use export_all=FALSE. This more closely simulates behavior when loading an installed package with library(), and can be useful for checking for missing exports.

Shim files

load_all also inserts shim functions into the imports environment of the loaded package. It presently adds a replacement version of system.file which returns different paths from base::system.file. This is needed because installed and uninstalled package sources have different directory structures. Note that this is not a perfect replacement for base::system.file.

Examples

# \dontrun{
# Load the package in the current directory
load_all("./")
#> Error in load_all("./"): could not find function "load_all"

# Running again loads changed files
load_all("./")
#> Error in load_all("./"): could not find function "load_all"

# With reset=TRUE, unload and reload the package for a clean start
load_all("./", TRUE)
#> Error in load_all("./", TRUE): could not find function "load_all"

# With export_all=FALSE, only objects listed as exports in NAMESPACE
# are exported
load_all("./", export_all = FALSE)
#> Error in load_all("./", export_all = FALSE): could not find function "load_all"
# }